backend/requirements.txt
Django==4.2.14
django-comment-migrate==0.1.7
django-cors-headers==4.4.0
django-filter==24.2
django-ranged-response==0.2.0
djangorestframework==3.15.2
django-restql==0.15.4
django-simple-captcha==0.6.0
django-timezone-field==7.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
mysqlclient==2.2.0
pypinyin==0.51.0
ua-parser==0.18.0
pyparsing==3.1.2
openpyxl==3.1.5
requests==2.32.3
typing-extensions==4.12.2
tzlocal==5.2
channels==4.1.0
channels-redis==4.2.0
websockets==11.0.3
user-agents==2.2.0
six==1.16.0
whitenoise==6.7.0
psycopg2==2.9.9
uvicorn==0.30.3
gunicorn==22.0.0
gevent==24.2.1
Pillow==10.4.0
pyinstaller==6.9.0
# celery
celery==5.4.0
eventlet==0.38.2
django-celery-beat>=2.6.0
tenant-schemas-celery>=3.0.0
django-redis>=5.4.0
django-celery-results>=2.5.1
# autoweb
pymysql>=1.1.1
selenium==4.27.1
